api-django-crud/pyproject.toml
[tool.poetry]
name = "api-django-crud"
version = "0.1.0"
description = ""
authors = ["Michel Clemer Monteiro Deolino <michelclemermonteirodeolino@gmial.com>"]
readme = "README.md"
packages = [{include = "api_django_crud"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
